backend/requirements.txt
# ProBharatAI Backend Dependencies
flask==3.1.0
flask-cors==5.0.1
flask-socketio==5.5.1
python-dotenv==1.1.0
playwright==1.50.0
pyautogui==0.9.54
openai==1.68.2
anthropic==0.49.0
google-generativeai==0.8.5
groq==0.18.0
httpx==0.28.1
langchain==0.3.20
langgraph==0.3.18
pydantic==2.10.6
python-telegram-bot==21.11
aiosqlite==0.21.0
sqlalchemy==2.0.38
alembic==1.15.1
celery==5.4.0
redis==5.2.1
requests==2.32.3
beautifulsoup4==4.13.3
pandas==2.2.3
openpyxl==3.1.5
gspread==6.1.5
click==8.1.8
rich==13.9.4
uvicorn==0.34.0
gunicorn==23.0.0
pyjwt==2.10.1
cryptography==44.0.1
psutil==6.1.1
